pyproject.toml
[tool.poetry]
name = "hashdotpy"
version = "0.1.0"
description = ""
authors = ["h"]
packages = [
    { include = "hash.py" },
]

[tool.poetry.dependencies]
python = "^3.8"
crcmod = "^1.7"
pycryptodome = "^3.10.1"
hashlib-additional = "^1.1"
blake3 = "^0.1.8"
siphash = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
run = "hash:init"